// src/utils/enum.ts
export enum Gender {
    MALE = 'Male',
    FEMALE = 'Female',
    OTHER = 'Other'
}

export enum Rol {
    MOTHER = 'Mother',
    FATHER = 'Father',
    SON = 'Son',
    DAUGHTER = 'Daughter',
    OTHER = 'Other'
}

export enum Gen {
    I = 'I',
    II = 'II',
    III = 'III',
    IV = 'IV',
    V = 'V'
}

export enum RelationTypeCouple {
    MATRIMONIO = 'Matrimonio',
    SEPARACION = 'Separación',
    DIVORCIO = 'Divorcio',
    CONVIVIENTES = 'Convivientes',
    OTHER = 'Other'
}

export enum RelationTypeMember {
    HIJO_LEGITIMO = 'HijoLegitimo',
    HIJO_ADOPTIVO = 'HijoAdoptivo',
    HIJO_LEGITIMO_GEMELO = 'HijoLegitimoGemelo',
    OTHER = 'Other'
}

export enum CurrentStatus {
    J = 'Jodido',
    C = 'Con ganas de vivir',
    E = 'En la mala',
    S = 'Solo Dios sabe',
    OTHER = 'Other'
}